+++ src/Product.java
@@ -1,9 +1,9 @@
 public class Product {
     public String name;
     public float price;
-    public short amount;
+    public int amount;
 
-    public Product(String name, float price, short amount) {
+    public Product(String name, float price, int amount) {
         this.name = name;
         this.price = price;
         this.amount = amount;
